Fix: add directives to simplify lifecycle and fix wrapper processes port management (#17)
This PR adds 2x directives that are set in the app-level databag:

* test_name: used to generate and track the current run of the benchmark
* stop: used to inform all the apps the need to stop
The later is relevant as we wanted to have the leader pushing all the peers for a stop of the application. This change simplifies the lifecycle management as the directives will be used to calculate where are we in the application stage.

Fixes copyright across the files. Updates everything to 2025.

This PR also standardizes the way the wrapper script manages the benchmark ports. The main problem is that we may have several benchmark scripts running in parallel and multiple ports may be open to manage each of these scripts. Therefore, we need to simplify the logic to manage these ports.